* p4a-own:
  Added option --retry-publish to p4a_pack.py that will publish all packages without rebuilding them. This option is to be used after a failure of par4all package publication.
  Added --publish-only option to p4a_pack to 'manually' copy/publish a file or several files. It's intended to be used after a failure of par4all package publication due to a network connection problem.

+++ src/simple_tools/p4a_pack.py
@@ -17,6 +17,7 @@ import optparse
 import platform
 import tempfile
 import shutil
+import glob
 
 '''
 Par4All packing routines: allows you to create .deb or .tar.gz packages of Par4all.
@@ -87,6 +88,12 @@ def add_module_options(parser):
 
     group.add_option("--publish", action = "store_true", default = False,
         help = "Publish the produced packages on the server.")
+        
+    group.add_option("--publish-only", action = "append", metavar = "FILE", default = [],
+        help = "Publish only a file (.deb or tgz or stgz) without rebuilding it. Several files are allowed")        
+
+    group.add_option("--retry-publish", action = "store_true",  default = False,
+        help = "Retry to publish only files (.deb and/or tgz and/or stgz) without rebuilding them.")        
 
     group.add_option("--release", dest = "development", action = "store_false", default = True,
         help = "When publishing, put the packages in release directories instead of development ones.")
@@ -232,20 +239,21 @@ def create_stgz(pack_dir, install_prefix, version, gitrev, keep_temp = False):
     return package_file
 
 
-def publish_deb(file, host, repos_dir, arch):
+def publish_deb(file, host, repos_dir, arch, publish_only=False):
     #~ arch = p4a_util.change_file_ext(file, "").split("_")[-1]
     p4a_util.info("Publishing " + file + " in the deb repository (" + arch + ")")
     repos_arch_dir = repos_dir + "/binary-" + arch
-    p4a_util.warn("Removing existing .deb file for arch " + arch + " (" + repos_arch_dir + ")")
-    p4a_util.run([ "ssh", host, "mkdir -p " + repos_arch_dir ])
-    p4a_util.run([ "ssh", host, "rm -fv " + repos_arch_dir + "/*.deb" ])
+    if not publish_only:
+		p4a_util.warn("Removing existing .deb file for arch " + arch + " (" + repos_arch_dir + ")")
+		p4a_util.run([ "ssh", host, "rm -fv " + repos_arch_dir + "/*.deb" ])
+    p4a_util.run([ "ssh", host, "mkdir -p " + repos_arch_dir ])    
     p4a_util.warn("Copying " + file + " on " + host + " (" + repos_arch_dir + ")")
-    p4a_util.run([ "rsync --partial --sparse", file, host + ":" + repos_arch_dir ])
+    p4a_util.run([ "rsync --partial --sparse --timeout=3600", file, host + ":" + repos_arch_dir ])
     p4a_util.warn("Please allow max. 5 minutes for deb repository indexes to get updated by cron")
     p4a_util.warn("Alternatively, you can run /srv/update-par4all.sh on " + host + " as root")
 
 
-def publish_files(files, distro, deb_distro, arch, deb_arch, development = False):
+def publish_files(files, distro, deb_distro, arch, deb_arch, development = False, publish_only=False):
     global default_publish_host
     global default_publish_dir, default_development_publish_dir
     global default_deb_publish_dir, default_deb_development_publish_dir
@@ -272,11 +280,11 @@ def publish_files(files, distro, deb_distro, arch, deb_arch, development = False
         p4a_util.warn("Copying " + file + " in " + publish_dir + " on " + default_publish_host)
         p4a_util.warn("If something goes wrong, try running /srv/fixacl-par4all.sh to fix permissions")
         p4a_util.warn("If something goes wrong, try creating directories or publishing the file manually")
-        p4a_util.run([ "ssh", default_publish_host, "mkdir -p " + publish_dir ])
-        p4a_util.run([ "rsync --partial --sparse", file, default_publish_host + ":" + publish_dir ])
+        p4a_util.run([ "ssh", default_publish_host, "mkdir -p " + publish_dir ])   
+        p4a_util.run([ "rsync --partial --sparse --timeout=3600", file, default_publish_host + ":" + publish_dir ])
         ext = p4a_util.get_file_ext(file)
         if ext == ".deb":
-            publish_deb(file, default_publish_host, deb_publish_dir, deb_arch)
+            publish_deb(file, default_publish_host, deb_publish_dir, deb_arch, publish_only)
 
 
 def work(options, args = []):
@@ -309,10 +317,14 @@ def work(options, args = []):
         if not options.publish:
             p4a_util.die("You specified files on command line but did not specify --publish")
         publish_files(args, distro, distro, arch, deb_arch, options.development)
-        return
+        return				
 
     if options.development and not options.append_date:
         options.append_date = True
+
+	if len(options.publish_only):
+		publish_files(options.publish_only, distro, distro, arch, deb_arch, options.development, publish_only=True)
+		return	
 
     if (not options.deb #and not options.sdeb
         and not options.tgz and not options.stgz):
@@ -321,6 +333,21 @@ def work(options, args = []):
         options.tgz = True
         options.stgz = True
 
+	if options.retry_publish:
+		file_to_publish=[]
+		if options.deb:
+			if options.pack_output_dir:
+				file_to_publish+=glob.glob(options.pack_output_dir+"/*.deb")
+			else:
+				file_to_publish+=glob.glob("*.deb")
+		if options.tgz or options.stgz:
+			if options.pack_output_dir:
+				file_to_publish+=glob.glob(options.pack_output_dir+"/*.gz")			
+			else:
+				file_to_publish+=glob.glob("*.gz")
+		publish_files(file_to_publish, distro, distro, arch, deb_arch, options.development, publish_only=True)
+		return	
+						
     prefix = options.install_prefix
     if prefix:
         p4a_util.warn("Installation prefix: " + prefix + " (--install-prefix)")
@@ -383,21 +410,28 @@ def work(options, args = []):
     if not p4a_util.which("fakeroot"):
         p4a_util.die("fakeroot not found, please install it (sudo aptitude install fakeroot)")
 
+
     global temp_dirs
     output_files = []
     try:
         if options.deb:
-            output_files.append(create_deb(pack_dir = pack_dir, install_prefix = prefix,
-                version = version, gitrev = gitrev, distro = distro, arch = deb_arch,
+			#remove previous *.deb
+			p4a_util.run(["rm -fv *.deb"])
+			output_files.append(create_deb(pack_dir = pack_dir, install_prefix = prefix,
+				version = version, gitrev = gitrev, distro = distro, arch = deb_arch,
                 keep_temp = options.keep_temp))
 
         if options.tgz:
-            output_files.append(create_tgz(pack_dir = pack_dir, install_prefix = prefix,
+			#remove previous *.gz
+			p4a_util.run(["rm -fv *.gz"])
+			output_files.append(create_tgz(pack_dir = pack_dir, install_prefix = prefix,
                 version = version, gitrev = gitrev, distro = distro, arch = arch,
                 keep_temp = options.keep_temp))
 
         if options.stgz:
-            output_files.append(create_stgz(pack_dir = pack_dir, install_prefix = prefix,
+			#remove previous *.gz
+			p4a_util.run(["rm -fv *src.tar.gz"])
+			output_files.append(create_stgz(pack_dir = pack_dir, install_prefix = prefix,
                 version = src_version, gitrev = src_gitrev,
                 keep_temp = options.keep_temp))
